requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
google-generativeai
python-dotenv
scikit-learn